event queries now batch multiple adjacent fights
if an event query only operates on a single boss and there are no
intervening trash fights, it will now require only a single request.
fingers crossed that we don't hit a boundary condition on what WCL is
willing to return. i tried deaths with no cutoff and it handled a bunch
of zaqul pulls without issue (225 records returned)